split files in s4571084
